Order status management

+++ client-service/src/main/java/com/example/client/service/service/convertor/ClientMapper.java
@@ -19,7 +19,7 @@ public interface ClientMapper {
     @Mapping(target = "id", ignore = true)
     void updateEntityToModel(@MappingTarget Client target, ClientDto source);
     @Mapping(target = "id", source = "id")
-    ClientDto requestToDto(Integer id, ClientRequest clientRequest);
+    ClientDto requestToDto(Long id, ClientRequest clientRequest);
 
     ClientDto requestToDto(ClientRequest clientRequest);
